Add DATABASE_SOCKET_DIR and related to README.md
Support for connecting to PostgreSQL via Unix domain socket was added a couple of years ago, and this should help advertise that support.

There are 3 additional influential variables that users should expect to need defining:
- DATABASE_NAME
- PGUSER
- PGPASSWORD

While DATABASE_NAME is Plausible-specific, PGUSER and PGPASSWORD support comes from the underlying database driver. Rather than making users dig through code and commit history for this info (because the driver doesn't document this either), we include it here for ease of access.